Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main(){
-    float f;
-    scanf("%f", &f);
+    double f;
+    scanf("%lf", &f);
     print("You entered: %.2f\n", f);
     return 0;
 }
